Moves OpenStreetMap extension to its own plugin
Complete with osmosis, models, eclipse plugin extensions, etc...
