+++ gama.dependencies/.classpath
@@ -1,6 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry excluding="gama/dependencies/kabeja/dxf/|gama/dependencies/kabeja/dxf/helpers/|gama/dependencies/kabeja/dxf/objects/|gama/dependencies/kabeja/io/|gama/dependencies/kabeja/math/|gama/dependencies/kabeja/parser/|gama/dependencies/kabeja/parser/dxf/|gama/dependencies/kabeja/parser/dxf/filter/|gama/dependencies/kabeja/parser/entities/|gama/dependencies/kabeja/parser/header/|gama/dependencies/kabeja/parser/objects/|gama/dependencies/kabeja/parser/table/|gama/dependencies/kabeja/processing/|gama/dependencies/kabeja/processing/event/|gama/dependencies/kabeja/processing/helper/|gama/dependencies/kabeja/tools/|gama/dependencies/kabeja/xml/" kind="src" path="src"/>
+	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jcc-12.1.0.0.jar"/>
+	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar"/>
+	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jaxb-runtime-4.0.7.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/gt-app-schema-33.4.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/gt-app-schema-resolver-33.4.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/gt-complex-33.4.jar"/>
@@ -10,8 +13,6 @@
 	<classpathentry exported="true" kind="lib" path="libs/apache-log4j-2.26.0/log4j-core-2.26.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/apache-log4j-2.26.0/log4j-slf4j2-impl-2.26.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/apache-log4j-2.26.0/log4j-slf4j-impl-2.26.0.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jaxb-runtime-4.0.7.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/jakarta.xml.bin-api 4.0.5/jakarta.xml.bind-api-4.0.5.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/commons-jxpath-1.3.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/jai-imageio-core 1.4.0/jai-imageio-core-1.4.0.jar"/>
@@ -26,7 +27,6 @@
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/net.opengis.wfs-33.4.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/joda-time-2.10.3.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jgridshift-core-1.3.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/jcc-12.1.0.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/gt-cql-33.4.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/commons-beanutils-1.11.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/commons-pool-1.5.4.jar"/>
@@ -44,23 +44,6 @@
 	<classpathentry exported="true" kind="lib" path="libs/jts 1.20.0/jts-core-1.20.0.jar" sourcepath="libs/jts 1.20.0/jts-core-1.20.0-sources.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/jts 1.20.0/jts-io-common-1.20.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/math3 3.6.1/commons-math3-3.6.1.jar" sourcepath="libs/math3 3.6.1/commons-math3-3.6.1-sources.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-0.49.2.jar" sourcepath="libs/osmosis 0.49.2/osmosis-core-0.49.2-sources.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-areafilter-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-core-0.49.2.jar" sourcepath="libs/osmosis 0.49.2/osmosis-core-0.49.2-sources.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-dataset-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-extract-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-hstore-jdbc-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-pbf-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-pbf2-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-pgsimple-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-pgsnapshot-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-replication-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-set-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-tagfilter-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-tagtransform-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmosis-xml-0.49.2.jar"/>
-	<classpathentry exported="true" kind="lib" path="libs/osmosis 0.49.2/osmpbf-1.6.0.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/streamex 0.8.4/streamex-0.8.4.jar" sourcepath="libs/streamex 0.8.4/streamex-0.8.4-sources.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/Java-WebSocket 1.6.0/Java-WebSocket-1.6.0.jar" sourcepath="libs/Java-WebSocket 1.6.0/Java-WebSocket-1.6.0-sources.jar"/>
 	<classpathentry exported="true" kind="lib" path="libs/geotools 33.4/bigint-0.7.1.jar"/>
