+++ gama.dependencies/META-INF/MANIFEST.MF
@@ -232,9 +232,6 @@ Export-Package: COM.ibm.db2.app,
  com.sun.media.jai.tilecodec,
  com.sun.media.jai.util,
  com.sun.media.jai.widget,
- crosby.binary,
- crosby.binary.file,
- crosby.binary.osmosis,
  de.micromata.opengis.kml.v_2_2_0,
  de.micromata.opengis.kml.v_2_2_0.annotations,
  de.micromata.opengis.kml.v_2_2_0.atom,
@@ -1401,74 +1398,6 @@ Export-Package: COM.ibm.db2.app,
  org.locationtech.jts.triangulate.quadedge,
  org.locationtech.jts.triangulate.tri,
  org.locationtech.jts.util,
- org.openstreetmap.osmosis.apidb,
- org.openstreetmap.osmosis.apidb.common,
- org.openstreetmap.osmosis.apidb.v0_6,
- org.openstreetmap.osmosis.apidb.v0_6.impl,
- org.openstreetmap.osmosis.areafilter,
- org.openstreetmap.osmosis.areafilter.common,
- org.openstreetmap.osmosis.areafilter.v0_6,
- org.openstreetmap.osmosis.core,
- org.openstreetmap.osmosis.core.bound.v0_6,
- org.openstreetmap.osmosis.core.buffer.v0_6,
- org.openstreetmap.osmosis.core.change.v0_6.impl,
- org.openstreetmap.osmosis.core.cli,
- org.openstreetmap.osmosis.core.container.v0_6,
- org.openstreetmap.osmosis.core.database,
- org.openstreetmap.osmosis.core.domain.common,
- org.openstreetmap.osmosis.core.domain.v0_6,
- org.openstreetmap.osmosis.core.filter.common,
- org.openstreetmap.osmosis.core.lifecycle,
- org.openstreetmap.osmosis.core.merge.common,
- org.openstreetmap.osmosis.core.misc.v0_6,
- org.openstreetmap.osmosis.core.pipeline.common,
- org.openstreetmap.osmosis.core.pipeline.v0_6,
- org.openstreetmap.osmosis.core.plugin,
- org.openstreetmap.osmosis.core.progress.v0_6,
- org.openstreetmap.osmosis.core.progress.v0_6.impl,
- org.openstreetmap.osmosis.core.report.v0_6,
- org.openstreetmap.osmosis.core.runner,
- org.openstreetmap.osmosis.core.sort.common,
- org.openstreetmap.osmosis.core.sort.v0_6,
- org.openstreetmap.osmosis.core.store,
- org.openstreetmap.osmosis.core.task.common,
- org.openstreetmap.osmosis.core.task.v0_6,
- org.openstreetmap.osmosis.core.tee.v0_6,
- org.openstreetmap.osmosis.core.time,
- org.openstreetmap.osmosis.core.util,
- org.openstreetmap.osmosis.dataset,
- org.openstreetmap.osmosis.dataset.v0_6,
- org.openstreetmap.osmosis.extract.apidb.common,
- org.openstreetmap.osmosis.extract.apidb.v0_6,
- org.openstreetmap.osmosis.hstore,
- org.openstreetmap.osmosis.pbf2,
- org.openstreetmap.osmosis.pbf2.v0_6,
- org.openstreetmap.osmosis.pbf2.v0_6.impl,
- org.openstreetmap.osmosis.pgsimple,
- org.openstreetmap.osmosis.pgsimple.common,
- org.openstreetmap.osmosis.pgsimple.v0_6,
- org.openstreetmap.osmosis.pgsimple.v0_6.impl,
- org.openstreetmap.osmosis.pgsnapshot,
- org.openstreetmap.osmosis.pgsnapshot.common,
- org.openstreetmap.osmosis.pgsnapshot.v0_6,
- org.openstreetmap.osmosis.pgsnapshot.v0_6.impl,
- org.openstreetmap.osmosis.replication,
- org.openstreetmap.osmosis.replication.common,
- org.openstreetmap.osmosis.replication.v0_6,
- org.openstreetmap.osmosis.replication.v0_6.impl,
- org.openstreetmap.osmosis.set,
- org.openstreetmap.osmosis.set.v0_6,
- org.openstreetmap.osmosis.set.v0_6.impl,
- org.openstreetmap.osmosis.tagfilter,
- org.openstreetmap.osmosis.tagfilter.common,
- org.openstreetmap.osmosis.tagfilter.v0_6,
- org.openstreetmap.osmosis.tagtransform,
- org.openstreetmap.osmosis.tagtransform.impl,
- org.openstreetmap.osmosis.tagtransform.v0_6,
- org.openstreetmap.osmosis.xml,
- org.openstreetmap.osmosis.xml.common,
- org.openstreetmap.osmosis.xml.v0_6,
- org.openstreetmap.osmosis.xml.v0_6.impl,
  org.picocontainer,
  org.picocontainer.alternatives,
  org.picocontainer.defaults,
@@ -1679,8 +1608,6 @@ Require-Bundle: gama.dev;visibility:=reexport,
  com.sun.jna.platform
 Bundle-ClassPath: .,
  libs/geotools 33.4/protobuf-java-3.25.5.jar,
- libs/osmosis 0.49.2/osmpbf-1.6.0.jar,
- libs/osmosis 0.49.2/osmosis-pbf-0.49.2.jar,
  libs/geotools 33.4/bigint-0.7.1.jar,
  libs/geotools 33.4/cdm-core-5.5.3.jar,
  libs/geotools 33.4/commons-dbcp-1.4.jar,
@@ -1812,26 +1739,10 @@ Bundle-ClassPath: .,
  libs/jts 1.20.0/jts-core-1.20.0.jar,
  libs/jts 1.20.0/jts-io-common-1.20.0.jar,
  libs/math3 3.6.1/commons-math3-3.6.1.jar,
- libs/osmosis 0.49.2/osmosis-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-areafilter-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-core-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-dataset-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-extract-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-hstore-jdbc-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-pbf2-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-pgsimple-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-pgsnapshot-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-replication-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-set-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-tagfilter-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-tagtransform-0.49.2.jar,
- libs/osmosis 0.49.2/osmosis-xml-0.49.2.jar,
  libs/streamex 0.8.4/streamex-0.8.4.jar,
  libs/geotools 33.4/commons-pool-1.5.4.jar,
  libs/geotools 33.4/commons-beanutils-1.11.0.jar,
  libs/geotools 33.4/gt-cql-33.4.jar,
- libs/geotools 33.4/jcc-12.1.0.0.jar,
  libs/geotools 33.4/jgridshift-core-1.3.jar,
  libs/geotools 33.4/joda-time-2.10.3.jar,
  libs/geotools 33.4/net.opengis.wfs-33.4.jar,
@@ -1846,8 +1757,6 @@ Bundle-ClassPath: .,
  libs/geotools 33.4/jai_imageio-1.1.jar,
  libs/geotools 33.4/commons-jxpath-1.3.jar,
  libs/jakarta.xml.bin-api 4.0.5/jakarta.xml.bind-api-4.0.5.jar,
- libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar,
- libs/geotools 33.4/jaxb-runtime-4.0.7.jar,
  libs/apache-log4j-2.26.0/log4j-1.2-api-2.26.0.jar,
  libs/apache-log4j-2.26.0/log4j-api-2.26.0.jar,
  libs/apache-log4j-2.26.0/log4j-core-2.26.0.jar,
@@ -1856,6 +1765,9 @@ Bundle-ClassPath: .,
  libs/geotools 33.4/jt-shadedrelief-1.1.31.jar,
  libs/geotools 33.4/gt-app-schema-33.4.jar,
  libs/geotools 33.4/gt-app-schema-resolver-33.4.jar,
- libs/geotools 33.4/gt-complex-33.4.jar
+ libs/geotools 33.4/gt-complex-33.4.jar,
+ libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar,
+ libs/geotools 33.4/jaxb-runtime-4.0.7.jar,
+ libs/geotools 33.4/jcc-12.1.0.0.jar
 Bundle-RequiredExecutionEnvironment: JavaSE-25
 Automatic-Module-Name: gama.dependencies
