+++ gama.dependencies/build.properties
@@ -4,8 +4,6 @@ bin.includes = .,\
                gama.feature.dependencies@2x.png,\
                about.ini,\
                libs/geotools 33.4/protobuf-java-3.25.5.jar,\
-               libs/osmosis 0.49.2/osmpbf-1.6.0.jar,\
-               libs/osmosis 0.49.2/osmosis-pbf-0.49.2.jar,\
                libs/geotools 33.4/bigint-0.7.1.jar,\
                libs/geotools 33.4/cdm-core-5.5.3.jar,\
                libs/geotools 33.4/commons-dbcp-1.4.jar,\
@@ -137,26 +135,10 @@ bin.includes = .,\
                libs/jts 1.20.0/jts-core-1.20.0.jar,\
                libs/jts 1.20.0/jts-io-common-1.20.0.jar,\
                libs/math3 3.6.1/commons-math3-3.6.1.jar,\
-               libs/osmosis 0.49.2/osmosis-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-areafilter-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-core-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-dataset-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-extract-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-hstore-jdbc-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-pbf2-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-pgsimple-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-pgsnapshot-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-replication-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-set-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-tagfilter-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-tagtransform-0.49.2.jar,\
-               libs/osmosis 0.49.2/osmosis-xml-0.49.2.jar,\
                libs/streamex 0.8.4/streamex-0.8.4.jar,\
                libs/geotools 33.4/commons-pool-1.5.4.jar,\
                libs/geotools 33.4/commons-beanutils-1.11.0.jar,\
                libs/geotools 33.4/gt-cql-33.4.jar,\
-               libs/geotools 33.4/jcc-12.1.0.0.jar,\
                libs/geotools 33.4/jgridshift-core-1.3.jar,\
                libs/geotools 33.4/joda-time-2.10.3.jar,\
                libs/geotools 33.4/net.opengis.wfs-33.4.jar,\
@@ -171,8 +153,6 @@ bin.includes = .,\
                libs/geotools 33.4/jai_imageio-1.1.jar,\
                libs/geotools 33.4/commons-jxpath-1.3.jar,\
                libs/jakarta.xml.bin-api 4.0.5/jakarta.xml.bind-api-4.0.5.jar,\
-               libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar,\
-               libs/geotools 33.4/jaxb-runtime-4.0.7.jar,\
                libs/apache-log4j-2.26.0/log4j-1.2-api-2.26.0.jar,\
                libs/apache-log4j-2.26.0/log4j-api-2.26.0.jar,\
                libs/apache-log4j-2.26.0/log4j-core-2.26.0.jar,\
@@ -181,7 +161,10 @@ bin.includes = .,\
                libs/geotools 33.4/jt-shadedrelief-1.1.31.jar,\
                libs/geotools 33.4/gt-app-schema-33.4.jar,\
                libs/geotools 33.4/gt-app-schema-resolver-33.4.jar,\
-               libs/geotools 33.4/gt-complex-33.4.jar
+               libs/geotools 33.4/gt-complex-33.4.jar,\
+               libs/geotools 33.4/jaxb-api-2.4.0-b180830.0359.jar,\
+               libs/geotools 33.4/jaxb-runtime-4.0.7.jar,\
+               libs/geotools 33.4/jcc-12.1.0.0.jar
 	source.. = src/
 	output.. = bin/
 bin.excludes = libs/Java-WebSocket 1.6.0/Java-WebSocket-1.6.0-sources.jar,\
