+++ gama.extension.osm/.factorypath
@@ -1,0 +1,3 @@
+<factorypath>
+    <factorypathentry kind="WKSPJAR" id="/gama.processor/gama.processor.jar" enabled="true" runInBatchMode="false"/>
+</factorypath>
