+++ gama.extension.osm/META-INF/MANIFEST.MF
@@ -1,0 +1,34 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Gama OpenStreetMap Extension
+Bundle-SymbolicName: gama.extension.osm;singleton:=true
+Bundle-Version: 0.0.0.qualifier
+Export-Package: gama.extension.osm
+Bundle-Vendor: UMMISCO
+Require-Bundle: gama.core,
+ gama.dependencies,
+ org.eclipse.core.resources;bundle-version="[3.23.0,4.0.0)",
+ org.eclipse.core.runtime;bundle-version="3.34.200"
+Bundle-ClassPath: .,
+ libs/osmosis 0.49.2/osmosis-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-areafilter-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-core-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-core-0.49.2-sources.jar,
+ libs/osmosis 0.49.2/osmosis-dataset-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-extract-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-hstore-jdbc-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-pbf-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-pbf2-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-pgsimple-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-pgsnapshot-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-replication-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-set-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-tagfilter-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-tagtransform-0.49.2.jar,
+ libs/osmosis 0.49.2/osmosis-xml-0.49.2.jar,
+ libs/osmosis 0.49.2/osmpbf-1.6.0.jar
+Bundle-ActivationPolicy: lazy
+Bundle-RequiredExecutionEnvironment: JavaSE-25
+Automatic-Module-Name: gama.extension.osm
+
