+++ gama.extension.osm/build.properties
@@ -1,0 +1,27 @@
+source.. = src/,\
+           gaml/
+output.. = bin/
+jre.compilation.profile = JavaSE-25
+bin.includes = .,\
+               plugin.xml,\
+               META-INF/,\
+               models/,\
+               libs/osmosis 0.49.2/osmosis-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-areafilter-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-core-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-core-0.49.2-sources.jar,\
+               libs/osmosis 0.49.2/osmosis-dataset-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-extract-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-hstore-jdbc-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-pbf-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-pbf2-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-pgsimple-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-pgsnapshot-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-replication-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-set-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-tagfilter-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-tagtransform-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmosis-xml-0.49.2.jar,\
+               libs/osmosis 0.49.2/osmpbf-1.6.0.jar
+javacDefaultEncoding.. = UTF-8
