+++ gama.extension.osm/libs/osmosis 0.49.2/osmosis-apidb-0.49.2.jar
