+++ gama.extension.osm/libs/osmosis 0.49.2/osmosis-core-0.49.2.jar
