+++ gama.extension.osm/models/OpenStreetMap/.project
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<projectDescription>
+	<name>OpenStreetMap</name>
+	<comment>osm plugin</comment>
+	<projects>
+	</projects>
+	<buildSpec>
+		<buildCommand>
+			<name>org.eclipse.xtext.ui.shared.xtextBuilder</name>
+			<arguments>
+			</arguments>
+		</buildCommand>
+	</buildSpec>
+	<natures>
+		<nature>org.eclipse.xtext.ui.shared.xtextNature</nature>
+		<nature>gama.workspace.gamaNature</nature>
+		<nature>gama.workspace.pluginNature</nature>
+	</natures>
+</projectDescription>
