+++ gama.extension.osm/models/OpenStreetMap/includes/boundary.shx
