+++ gama.extension.osm/models/OpenStreetMap/models/Download Data (OSM-images).gaml
