+++ gama.extension.osm/models/OpenStreetMap/models/OSM File Import.gaml
