+++ gama.parent/pom.xml
@@ -135,6 +135,8 @@
 		<module>../gama.extension.bdi</module>
 		<module>../gama.extension.batch</module>
 		<module>../gama.extension.fipa</module>
+		<module>../gama.extension.osm</module>
+		<module>../gama.extension.dataframe</module>
 		<module>../gama.extension.sound</module>
 		<module>../gama.extension.stats</module>
 		<module>../gama.extension.maths</module>
