+++ gama.ui.navigator/META-INF/MANIFEST.MF
@@ -20,7 +20,8 @@ Require-Bundle: gama.ui.shared,
  org.eclipse.ltk.core.refactoring,
  org.eclipse.ltk.ui.refactoring,
  gama.workspace,
- gaml.compiler
+ gaml.compiler,
+ gama.extension.osm;bundle-version="[0.0.0,1.0.0)"
 Bundle-ActivationPolicy: lazy
 Export-Package: gama.ui.navigator.commands,
  gama.ui.navigator.view,
