+++ gama.ui.navigator/src/gama/ui/navigator/view/contents/WrappedFile.java
@@ -29,8 +29,8 @@ import gama.api.kernel.GamaMetaModel;
 import gama.api.utils.files.IFileMetadataProvider;
 import gama.api.utils.files.IGamaFileMetaData;
 import gama.core.util.file.GMLInfo;
-import gama.core.util.file.OSMInfo;
 import gama.core.util.file.ShapeInfo;
+import gama.extension.osm.OSMInfo;
 import gama.ui.navigator.view.NavigatorContentProvider;
 import gama.ui.shared.resources.GamaColors;
 import gama.ui.shared.resources.GamaIcon;
